% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{decktape}
\alias{decktape}
\title{Converting xaringan HTML slides to PDF}
\usage{
decktape(xaringan_path, pdf_path, decktape_version = NULL,
  open_pdf = FALSE)
}
\arguments{
\item{xaringan_path}{The path to the xaringan HTML slides.}

\item{pdf_path}{The desired output path of the PDF file.}

\item{decktape_version}{Which `decktape` version should be used. Default is
to use the latest one.}

\item{open_pdf}{Should the resulting PDF be opened with your system PDF
viewer?}
}
\description{
This function uses the hosted docker image of the nodejs library 'decktape'
to convert xaringan HTML slides (which are based on 'remark.js' under
the hood) to PDF.
}
\section{Docker}{

To run this function you need to have a working installation of
\href{https://www.docker.com/}{docker}. For some operating systems you may need
to \href{https://stackoverflow.com/questions/48957195}{add yourself to the "docker" group}
and restart your machine.
}

\section{Decktape}{

This function uses the nodejs library \href{(https://github.com/astefanutti/decktape}{decktape}.
In fact the hosted docker version of decktape is used here.
By default the latest decktape version is used.
In case of errors you may want to try older decktape versions (version
2.8.0 works fine for example) using the `decktape_version` argument.
}

\section{Open-PDF}{

The opening of PDF relies on two points: Your operating system and your
PDF viewer. Some setups work, some do not. It is recommended to use full paths
as the internal function used does not like paths like "~/file.pdf".
}

