% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ninjutsu.R
\name{shadow_clone}
\alias{shadow_clone}
\title{Use a new xaringan theme.}
\usage{
shadow_clone(theme = NULL, ref = "master")
}
\arguments{
\item{theme}{Either a built-in theme shipped with xaringan, or the repository address of the theme in the format of username/repo}

\item{ref}{desired git reference. could be a commit SHA, or a branch name. Defauts to master}
}
\value{
The cloned theme
}
\description{
This function clones an xaringan theme for you to use. Include it in a code chunk in the xaringan slide source.
}
\examples{
# built-in
shadow_clone("kunoichi")
# remote
shadow_clone("tcgriffith/xaringan_theme_example")

}
