% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{infinite_moon_reader}
\alias{infinite_moon_reader}
\alias{inf_mr}
\title{Serve and live reload slides}
\usage{
infinite_moon_reader(moon, cast_from = ".", ...)

inf_mr(moon, cast_from = ".", ...)
}
\arguments{
\item{moon}{The input Rmd file path (if missing and in RStudio, the current
active document is used).}

\item{cast_from}{The root directory of the server.}

\item{...}{Passed to \code{rmarkdown::\link[rmarkdown]{render}()}.}
}
\description{
Use the \pkg{servr} package to serve and reload slides on change.
\code{inf_mr()} is an alias of \code{infinite_moon_reader()}.
}
\details{
The Rmd document is compiled continuously to trap the world in the Infinite
Tsukuyomi. The genjutsu is cast from the directory specified by
\code{cast_from}, and the Rinne Sharingan will be reflected off of the
\code{moon}.
}
\note{
This function is not really tied to the output format
  \code{\link{moon_reader}()}. You can use it to serve any single-HTML-file R
  Markdown output.
}
\references{
\url{https://naruto.fandom.com/wiki/Infinite_Tsukuyomi}
}
\seealso{
\code{servr::\link{httw}}
}
